+++ raiko-host/Cargo.toml
@@ -3,10 +3,6 @@ name = "raiko-host"
 version = "0.1.0"
 edition = "2021"
 
-[lib]
-name = "shit"
-path = "src/shit.rs"
-
 [dependencies]
 anyhow = "1.0"
 alloy-rlp = { version = "0.3", default-features = false }
@@ -47,10 +43,7 @@ lazy_static = "1.4.0"
 once_cell = "1.8.0"
 thiserror = "1.0"
 # powdr = { git = "https://github.com/ceciliaz030/powdr", branch = "+nightly", optional = true}
-# sp1-core = { git = "https://github.com/succinctlabs/sp1.git", optional = true }
 sp1-guest = { path = "../raiko-guests/succinct/script", optional = true}
-# risc0-zkvm = { version = "0.20.1", features = ["prove", "getrandom"], optional = true }
-# bonsai-sdk = { version = "0.6.0", features = ["async"],  optional = true }
 risc0-guest = { path = "../raiko-guests/risc0", optional = true}
 log = {version = "0.4"}
 typetag ={version = "0.2.15", optional = true}
@@ -73,7 +66,7 @@ rstest = "0.18"
 sp1-helper = { git = "https://github.com/succinctlabs/sp1.git" }
 
 [features]
-default = ["risc0"]
+default = []
 # powdr = ["dep:powdr"]
 succinct = ["dep:sp1-guest"]
 risc0 = [
