--- raiko-host/src/prover/consts.rs
@@ -1,5 +1,0 @@
-#[allow(dead_code)]
-pub const SGX_PARENT_DIR: &str = "sgx";
-#[allow(dead_code)]
-pub const PSE_ZK_PARENT_DIR: &str = "pse-zk";
-pub const RAIKO_GUEST_EXECUTABLE: &str = "raiko-guest";
