+++ raiko-host/src/prover/mod.rs
@@ -1,11 +1,8 @@
-pub mod consts;
 pub mod context;
+pub mod cache;
 pub mod error;
-// pub mod execution;
-pub mod execution_;
-
+pub mod execution;
 pub mod json_rpc;
-pub mod proof;
 pub mod request;
 pub mod server;
 pub mod utils;
