--- raiko-host/src/prover/proof/pse_zk.rs
@@ -1,1 +1,0 @@
-//! Pse zkevm proof
