+++ raiko-host/src/prover/utils.rs
@@ -1,6 +1,7 @@
 use std::path::{Path, PathBuf};
 
-use super::consts::RAIKO_GUEST_EXECUTABLE;
+pub const RAIKO_GUEST_EXECUTABLE: &str = "raiko-guest";
+
 
 pub fn cache_file_path(cache_path: &Path, block_no: u64, is_l1: bool) -> PathBuf {
     let prefix = if is_l1 { "l1" } else { "l2" };
