--- raiko-host/src/shit.rs
@@ -1,1 +1,0 @@
-pub struct Shit {}
